Misc cleanups.

+++ include/info.h
@@ -36,8 +36,8 @@
 char *Info_ValueForKey (char *s, char *key);
 void Info_RemoveKey (char *s, char *key);
 void Info_RemovePrefixedKeys (char *start, char prefix);
-void Info_SetValueForKey (char *s, char *key, char *value, int maxsize);
-void Info_SetValueForStarKey (char *s, char *key, char *value, int maxsize);
+void Info_SetValueForKey (char *s, char *key, char *value, size_t maxsize);
+void Info_SetValueForStarKey (char *s, char *key, char *value, size_t maxsize);
 void Info_Print (char *s);
 
 #endif	// _INFO_H
