+++ source/info.c
@@ -190,7 +190,7 @@ void Info_RemovePrefixedKeys (char *start, char prefix)
 }
 
 
-void Info_SetValueForStarKey (char *s, char *key, char *value, int maxsize)
+void Info_SetValueForStarKey (char *s, char *key, char *value, size_t maxsize)
 {
 	char	newstr[1024], *v;
 	int		c, is_name, is_team;
@@ -257,7 +257,7 @@ void Info_SetValueForStarKey (char *s, char *key, char *value, int maxsize)
 	*s = 0;
 }
 
-void Info_SetValueForKey (char *s, char *key, char *value, int maxsize)
+void Info_SetValueForKey (char *s, char *key, char *value, size_t maxsize)
 {
 	if (key[0] == '*')
 	{
